Show file counts in progressbar dialog when emptying trash

+++ libcore/FileOperations/EmptyTrashJob.vala
@@ -18,6 +18,9 @@
 
 public class Files.FileOperations.EmptyTrashJob : CommonJob {
     private GLib.List<GLib.File> trash_dirs;
+    private int total_files;
+    private int files_left;
+    private bool count_only;
 
     public EmptyTrashJob (Gtk.Window? parent_window = null, owned GLib.List<GLib.File>? trash_dirs = null) {
         base (parent_window);
@@ -46,6 +49,10 @@ public class Files.FileOperations.EmptyTrashJob : CommonJob {
                 var infos = yield enumerator.next_files_async (10, GLib.Priority.DEFAULT, cancellable);
                 while (infos.nth_data (0) != null) {
                     foreach (unowned GLib.FileInfo info in infos) {
+                        if (count_only) {
+                            total_files++;
+                            files_left++;
+                        }
                         var child = file.get_child (info.get_name ());
                         yield delete_trash_file (child, true, info.get_file_type () == GLib.FileType.DIRECTORY);
                     }
@@ -62,10 +69,11 @@ public class Files.FileOperations.EmptyTrashJob : CommonJob {
             return;
         }
 
-        if (delete_file) {
+        if (delete_file && !count_only) {
             try {
                 yield file.delete_async (GLib.Priority.DEFAULT, cancellable);
                 report_empty_trash_progress ();
+                files_left--;
             } catch (GLib.Error e) {
                 debug (e.message);
                 return;
@@ -120,17 +128,36 @@ public class Files.FileOperations.EmptyTrashJob : CommonJob {
 
     private void report_empty_trash_progress () {
         progress.take_status (_("Emptying trash"));
-        progress.take_details ("");
-        progress.pulse_progress ();
+
+        progress.take_status (ngettext (
+            "Emptying %'d file from trash",
+            "Emptying %'d files from trash",
+            total_files
+        ).printf (total_files));
+
+        progress.take_details (ngettext (
+            "%'d file left to trash",
+            "%'d files left to trash",
+            files_left
+        ).printf (files_left));
+
+        if (total_files != 0) {
+            progress.update_progress (total_files - files_left, total_files);
+        }
     }
 
     private async void internal_empty_trash () {
+        total_files = 0;
+        files_left = 0;
         progress.start ();
         foreach (unowned GLib.File dir in trash_dirs) {
             if (aborted ()) {
                 break;
             }
 
+            count_only = true;
+            yield delete_trash_file (dir, false, true);
+            count_only = false;
             yield delete_trash_file (dir, false, true);
         }
